# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hypermetrics API
  version: 1.0.0
  summary: Hyperliquid market and wallet data for builders
  description: |
    Query whale activity, wallet positions and aggregate Hyperliquid market data.

    ## Authentication

    Every request requires an active API key in the `HM-API-KEY` header. Create up
    to three keys from the Hypermetrics dashboard. A newly created secret is shown
    only once and can be revoked independently.

    ## Response format

    Successful responses use code `0`. Errors include a stable application code
    and a human-readable message. Timestamps are Unix milliseconds in UTC.
  contact:
    name: Hypermetrics support
    url: https://hypermetrics.xyz
  license:
    name: Proprietary
    url: https://hypermetrics.xyz
servers:
  - url: https://api.hypermetrics.xyz
    description: Configured API server
security:
  - ApiKeyAuth: []
...
paths:
  /api/hyperliquid/whale-alert:
    get:
      operationId: listWhaleAlerts
      tags: [Whales]
      summary: List recent whale alerts
      description: Returns the latest detected whale position events, capped at 200 results.
      responses:
        "200":
          description: Whale alerts loaded successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhaleAlertsResponse"
              example:
                code: "0"
                msg: success
                data:
                  - user: "0x1234567890abcdef1234567890abcdef12345678"
                    symbol: BTC
                    position_size: 4.25
                    entry_price: 114250.5
                    liq_price: 92110.2
                    position_value_usd: 487901.44
                    position_action: 1
                    create_time: 1785960000000
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
      x-codeSamples:
        - lang: Shell
          label: cURL
          source: |
            curl --request GET \
              --url 'https://api.hypermetrics.xyz/api/hyperliquid/whale-alert' \
              --header 'HM-API-KEY: hm_live_your_key'
  /api/hyperliquid/whale-position:
    get:
      operationId: listWhalePositions
      tags: [Whales]
      summary: List current whale positions
      description: Returns open positions above the whale threshold configured by Hypermetrics.
      responses:
        "200":
          description: Whale positions loaded successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhalePositionsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
      x-codeSamples:
        - lang: Shell
          label: cURL
          source: |
            curl --request GET \
              --url 'https://api.hypermetrics.xyz/api/hyperliquid/whale-position' \
              --header 'HM-API-KEY: hm_live_your_key'
  /api/hyperliquid/position:
    get:
      operationId: listPositionsBySymbol
      tags: [Positions]
      summary: List positions by market
      description: Returns a page of wallet positions for one Hyperliquid symbol. Each page contains up to 100 positions.
      parameters:
        - $ref: "#/components/parameters/Symbol"
        - $ref: "#/components/parameters/CurrentPage"
      responses:
        "200":
          description: Positions loaded successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
      x-codeSamples:
        - lang: Shell
          label: cURL
          source: |
            curl --request GET \
              --url 'https://api.hypermetrics.xyz/api/hyperliquid/position?symbol=BTC&current_page=1' \
              --header 'HM-API-KEY: hm_live_your_key'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: HM-API-KEY
      description: API key created from the Hypermetrics dashboard.
  parameters:
    Symbol:
      name: symbol
      in: query
      required: true
      description: Hyperliquid market symbol. Values are normalized to uppercase.
      schema:
        type: string
        minLength: 1
      example: BTC
    CurrentPage:
      name: current_page
      in: query
      required: false
      description: One-based page number.
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
  responses:
    BadRequest:
      description: A required query parameter is missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missingParameter:
              summary: Missing required parameter
              value:
                code: "1001"
                msg: "missing required parameter: symbol"
                data: null
            invalidParameter:
              summary: Invalid parameter
              value:
                code: "1002"
                msg: "invalid parameter: current_page"
                data: null
    Unauthorized:
      description: The API key is missing, invalid or revoked.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missingKey:
              summary: Missing API key
              value:
                code: "1000"
                msg: missing api key
                data: null
            invalidKey:
              summary: Invalid API key
              value:
                code: "1000"
                msg: invalid api key
                data: null
...
    InternalError:
      description: Hypermetrics could not complete the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: "1006"
            msg: failed to load data
            data: null
  schemas:
    SuccessCode:
      type: string
      const: "0"
      description: Stable application-level success code.
    ErrorResponse:
      type: object
      required: [code, msg, data]
      properties:
        code:
          type: string
          enum: ["1000", "1001", "1002", "1005", "1006"]
        msg:
          type: string
        data:
          type: "null"
    WhaleAlertsResponse:
      type: object
      required: [code, msg, data]
      properties:
        code:
          $ref: "#/components/schemas/SuccessCode"
        msg:
          type: string
          const: success
        data:
          type: array
          items:
            $ref: "#/components/schemas/WhaleAlert"
    WhalePositionsResponse:
      type: object
      required: [code, msg, data]
      properties:
        code:
          $ref: "#/components/schemas/SuccessCode"
        msg:
          type: string
          const: success
        data:
          type: array
          items:
            $ref: "#/components/schemas/WhalePosition"
    PositionsResponse:
      type: object
      required: [code, msg, data]
      properties:
        code:
          $ref: "#/components/schemas/SuccessCode"
        msg:
          type: string
          const: success
        data:
          type: array
          items:
            $ref: "#/components/schemas/Position"
...
    WhaleAlert:
      type: object
      required: [user, symbol, position_size, entry_price, liq_price, position_value_usd, position_action, create_time]
      properties:
        user:
          $ref: "#/components/schemas/Address"
        symbol:
          type: string
          example: BTC
        position_size:
          type: number
          format: double
          description: Signed asset quantity; positive is long and negative is short.
          example: 4.25
        entry_price:
          $ref: "#/components/schemas/UsdAmount"
        liq_price:
          $ref: "#/components/schemas/UsdAmount"
        position_value_usd:
          $ref: "#/components/schemas/UsdAmount"
        position_action:
          type: integer
          format: int32
          description: Internal position event action code.
          example: 1
        create_time:
          $ref: "#/components/schemas/UnixTimestampMilliseconds"
    WhalePosition:
      allOf:
        - $ref: "#/components/schemas/Position"
        - type: object
          required: [margin_balance]
          properties:
            margin_balance:
              $ref: "#/components/schemas/UsdAmount"
    Position:
      type: object
      required: [user, symbol, position_size, entry_price, mark_price, liq_price, leverage, margin_balance, position_value_usd, unrealized_pnl]
      properties:
        user:
          $ref: "#/components/schemas/Address"
        symbol:
          type: string
          example: BTC
        position_size:
          type: number
          format: double
          description: Signed asset quantity; positive is long and negative is short.
          example: 4.25
        entry_price:
          $ref: "#/components/schemas/UsdAmount"
        mark_price:
          $ref: "#/components/schemas/UsdAmount"
        liq_price:
          $ref: "#/components/schemas/UsdAmount"
        leverage:
          type: number
          format: double
          example: 5
        margin_balance:
          $ref: "#/components/schemas/UsdAmount"
        position_value_usd:
          $ref: "#/components/schemas/UsdAmount"
        unrealized_pnl:
          type: number
          format: double
          example: 18250.75
...
    Address:
      type: string
      pattern: "^0x[a-fA-F0-9]{40}$"
      example: "0x1234567890abcdef1234567890abcdef12345678"
    UsdAmount:
      type: number
      format: double
      description: USD-denominated value.
      example: 1250000.5
    UnixTimestampMilliseconds:
      type: integer
      format: int64
      description: Unix timestamp in milliseconds, UTC.
      example: 1785960000000